requirements.txt
streamlit
langchain
langgraph
langchain-community
ollama
openai
pytz
requests
python-dotenv
webvtt-py
python-docx
fpdf
langchain-ollama
groq>=0.4.2